curvefs-mds/curvefs-metaserver: install SIGTERM handler
Leader observation is handled in another detached thread,
and when observe leader failed, it raises a SIGTERM in the detached thread.

By default, we didn't register handler for SIGTERM, so the process should exit directly.
But, in current scenario, it seems the SIGTERM signal is ignored.

To avoid this problem, register handler for SIGTERM through brpc's flags graceful_quit_on_sigterm