curvefs:modify mds.conf
1. The original patition format defaults to 3, which is too small. This
will cause the metadata of an fs to only be distributed in 3 copies. Now
change this number to be consistent with the initial copy of the
cluster, which can make the initial distribution of fs metadata more
uniform.
2. Reduced the number of inodes in the partition, and let fs use as many
partitions as possible to make the metadata distribution more balanced.
3. Reduced the number of partitions in the copy to avoid managing too
much metadata in one copy.
4. The default number of copies in the cluster is adjusted to a multiple
of 3.