Fix side effect of Remove operation.
In the original ARC paper, there was no explanation of how Remove should be
handled when the cache is full. The Remove operation would have a side effect
on subsequent Put operations when the cache is not full, which is still
evicting items even though it's clear that there is no need to evict when
the cache is not full.

Now, code has been added to check if the cache is full to avoid performance
degradation.

- fix c_, it should be max_count.
- add unittest.
- use list::splice to avoid copying value
- add suffix to member fields list and map

Signed-off-by: Xu Yifeng <bsddiy2@163.com>